move bid wall tests to use launchpad_v7
