+++ programs/bid_wall/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "bid_wall"
-version = "0.6.0"
-description = "Created with Anchor"
+version = "0.7.0"
+description = "A big wall of cash sitting at NAV, ready to take tokens from weak hands and burn them."
 edition = "2021"
 
 [lib]
