+++ sdk/src/v0.6/index.ts
@@ -5,4 +5,3 @@ export * from "./FutarchyClient.js";
 export * from "./ConditionalVaultClient.js";
 export * from "./LaunchpadClient.js";
 export * from "./PriceBasedPerformancePackageClient.js";
-export * from "../v0.7/BidWallClient.js";
