+++ sdk/src/v0.7/constants.ts
@@ -20,6 +20,9 @@ export const SHARED_LIQUIDITY_MANAGER_PROGRAM_ID = new PublicKey(
 export const PRICE_BASED_PERFORMANCE_PACKAGE_PROGRAM_ID = new PublicKey(
   "pbPPQH7jyKoSLu8QYs3rSY3YkDRXEBojKbTgnUg7NDS",
 );
+export const BID_WALL_PROGRAM_ID = new PublicKey(
+  "WALL8ucBuUyL46QYxwYJjidaFYhdvxUFrgvBxPshERx",
+);
 
 export const MPL_TOKEN_METADATA_PROGRAM_ID = new PublicKey(
   "metaqbxxUerdq28cj1RbAWkYQm3ybzjb6a8bt518x1s",
