+++ tests/bidWall/unit/initializeBidWall.test.ts
@@ -11,7 +11,7 @@ import {
   BidWallClient,
   MAINNET_USDC,
   getBidWallAddr,
-} from "@metadaoproject/futarchy/v0.6";
+} from "@metadaoproject/futarchy/v0.7";
 import BN from "bn.js";
 import { getAssociatedTokenAddressSync } from "@solana/spl-token";
 import { initializeMintWithSeeds } from "../utils.js";
@@ -33,14 +33,14 @@ export default function suite() {
 
   before(async function () {
     futarchyClient = this.futarchy;
-    launchpadClient = this.launchpad_v6;
+    launchpadClient = this.launchpad_v7;
     bidWallClient = this.bidWall;
   });
 
   beforeEach(async function () {
     const result = await initializeMintWithSeeds(
       this.banksClient,
-      this.launchpad_v6,
+      this.launchpad_v7,
       this.payer,
     );
 
@@ -61,7 +61,7 @@ export default function suite() {
     );
 
     // Initialize launch
-    await this.launchpad_v6
+    await this.launchpad_v7
       .initializeLaunchIx({
         tokenName: "META",
         tokenSymbol: "META",
@@ -91,12 +91,20 @@ export default function suite() {
     // Advance clock and complete launch
     await this.advanceBySeconds(60 * 60 * 24 * 7 + 100);
     await launchpadClient.closeLaunchIx({ launch }).rpc();
+
+    await launchpadClient
+      .setFundingRecordApprovalIx({
+        launch,
+        funder: this.payer.publicKey,
+        approvedAmount: fundAmount,
+      })
+      .rpc();
+
     const completeLaunchTx = await launchpadClient
       .completeLaunchIx({
         launch,
         quoteMint: MAINNET_USDC,
         baseMint: META,
-        finalRaiseAmount: null,
         launchAuthority: this.payer.publicKey,
       })
       .transaction();
@@ -118,7 +126,7 @@ export default function suite() {
     await this.banksClient.processTransaction(tx);
 
     // Verify launch completion and DAO creation
-    const launchAccount = await this.launchpad_v6.fetchLaunch(launch);
+    const launchAccount = await this.launchpad_v7.fetchLaunch(launch);
     assert.exists(launchAccount.state.complete);
     assert.exists(launchAccount.dao);
     dao = launchAccount.dao;
@@ -135,7 +143,7 @@ export default function suite() {
     const feeRecipient = Keypair.generate().publicKey;
     await this.createTokenAccount(MAINNET_USDC, feeRecipient);
 
-    let launchAccount = await this.launchpad_v6.fetchLaunch(launch);
+    let launchAccount = await this.launchpad_v7.fetchLaunch(launch);
 
     await bidWallClient
       .initializeBidWallIx({
@@ -165,8 +173,6 @@ export default function suite() {
     const bidWallAccount = await bidWallClient.fetchBidWall(bidWall);
 
     assert.isNotNull(bidWallAccount);
-
-    console.log(bidWallAccount);
 
     assert.equal(
       bidWallAccount.authority.toBase58(),
