+++ tests/bidWall/utils.ts
@@ -1,11 +1,11 @@
 import { PublicKey, Signer, SystemProgram, Transaction } from "@solana/web3.js";
 import * as token from "@solana/spl-token";
 import { BanksClient } from "solana-bankrun";
-import { LaunchpadClient } from "@metadaoproject/futarchy/v0.6";
+import { LaunchpadClient } from "@metadaoproject/futarchy/v0.7";
 import {
   getLaunchAddr,
   getLaunchSignerAddr,
-} from "@metadaoproject/futarchy/v0.6";
+} from "@metadaoproject/futarchy/v0.7";
 
 export async function initializeMintWithSeeds(
   banksClient: BanksClient,
