+++ tests/main.test.ts
@@ -735,7 +735,7 @@ describe("launchpad_v7", launchpad_v7);
 describe("price_based_performance_package", priceBasedPerformancePackage);
 describe("conditional_vault", conditionalVault);
 describe("futarchy", futarchy);
-describe.only("bid_wall", bidWall);
+describe("bid_wall", bidWall);
 describe("project-wide integration tests", function () {
   it.skip("mint and swap in a single transaction", mintAndSwap);
   describe("full launch v6", fullLaunch);
